This implementation works, i changed all my functions into files to better organization
