+++ app.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "functions.h"
+#include "helpers.h"
 #include "classes.h"
 
 int main(void) {
@@ -10,7 +10,8 @@ int main(void) {
     bienvenida();
 
     while(true) {
-        std::vector<std::string> text = {"Login", "Register"};
+        std::vector<std::string> text;
+        text = {"Login", "Register"};
         showText(text);
         std::cin >> option;
         std::vector <std::string> credentials;
@@ -24,8 +25,7 @@ int main(void) {
             }
             break;
         case 2:
-            store.registrarse() ? std::cout << "Registrado Correctamente" : 
-            std::cout << "Registro invalido" << std::endl;
+            store.registrarse();
 
             break;
         
@@ -35,23 +35,4 @@ int main(void) {
         }
     }
 
-    // while(running) {
-    //     opcionesMain();
-    //     std::cout << "Type -1 to quit" << std::endl;
-    //     std::cin >> option;
-    //     switch (option) {
-    //     case -1:
-    //         running = false;
-    //         break;
-    //     case 1:
-    //         break;
-        
-    //     default:
-    //         std::cout << "Not valid" << std::endl;
-    //         break;
-    //     }
-        
-
-    // }
-    // store.showClients();
 }
