+++ helpers.cpp
@@ -1,14 +1,21 @@
-#ifndef FUNCTIONS_H
-#define FUNCTIONS_H
+#ifndef HELPERS_H
+#define HELPERS_H
 
-#include <vector>
 #include <iostream>
-#include "classes.h"
+#include "helpers.h"
+
+void delimiter(std::string _, int len) {
+    for (int i = 0; i < len; i++) {
+        std::cout << _;
+    }
+    std::cout << std::endl;
+}
+
 
 void bienvenida() {
     std::cout << "******** " <<"Bienvenido a la libreria de libros!" << " ********" << std::endl;
     std::cout << "Donde solo vendemos libros y mas libros y otros cuantos libros" << std::endl;
-    std::cout << "---------------------------------------------------------------------------------\n";
+    delimiter("-", 50);
     std::cout << "Para interactuar con el menu, simplemente escrbe el numero del menu al que quieres acceder" << std::endl;
 }
 
@@ -19,18 +26,18 @@ void showText(std::vector<std::string> text) {
     }
 }
 
-// std::vector<std::string> getCredentials() {
-//     std::vector<std::string> credentials;
-//     std::string username;
-//     std::string password;
-//     std::cout << "Escribe tu nombre de usuario: ";
-//     std::cin >> username;
-//     std::cout << "\nEscribe tu contraseña: " << std::endl;
-//     std::cin >> password;
-//     credentials.push_back(username);
-//     credentials.push_back(password);
-//     return credentials;
-// }
+std::vector<std::string> getCredentials() {
+    std::vector<std::string> credentials;
+    std::string username;
+    std::string password;
+    std::cout << "Escribe tu nombre de usuario: ";
+    std::cin >> username;
+    std::cout << "\nEscribe tu contraseña: " << std::endl;
+    std::cin >> password;
+    credentials.push_back(username);
+    credentials.push_back(password);
+    return credentials;
+}
 
 void opcionesMain() {
     std::vector<std::string> text = {"Iniciar Sesion"};
@@ -42,5 +49,6 @@ void opcionesCarrito() {
     "Ver subtotal", "Ver libros comprados", "Logout"};
     showText(text);
 }
+
 
 #endif
